bot/requirements.txt
dotenv-python==0.0.1
telepot==12.0
